training: disable fp16 (bf16/fp32 only) — MoE PolyGLU fp16 overflow under Muon
FP16's 5-bit exponent overflows inside the PolyGLU experts once Muon (grad-independent
updates) grows the weights, giving a sudden forward NaN with no gradient blow-up. bf16
shares fp32's exponent range and is stable. AMP_DTYPE now resolves from BIBO_AMP_DTYPE=
bf16(default)|fp32 and hard-errors on fp16; GradScaler gated to fp16; eval/compare autocast -> bf16.

+++ legacy/bench/compare.py
@@ -49,7 +49,7 @@ def per_seq_bpb(model, loader, tokenizer, device, max_batches):
         if nb >= max_batches:
             break
         ids = batch["input_ids"].to(device); labels = batch["labels"].to(device)
-        with torch.autocast("cuda", dtype=torch.float16):
+        with torch.autocast("cuda", dtype=torch.bfloat16):
             logits = model(input_ids=ids).logits              # (B,S,V) — labels omitted ⇒ logits returned
         B, S, V = logits.shape
         nll = F.cross_entropy(logits.float().view(-1, V), labels.view(-1),
